Added changes except for inferring the region
